tools/requirements.txt
cobs
protobuf
pyserial
pyusb
